Split color.c into separate files for better organization
- Split color.c into three files: color.c, cycle.c, and cycleset.c
- Created corresponding header files: color.h, cycle.h, and cycleset.h
- Updated Makefile to include the new source files
- Fixed include dependencies in existing files
- Maintained code functionality with proper type definitions

This refactoring improves maintainability by separating concerns:
- color.c: Only contains color and colorset operations
- cycle.c: Implements cycle operations
- cycleset.c: Implements cycleset operations
